+++ CHANGELOG.md
@@ -4,6 +4,19 @@ Change Log
 
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
+
+[0.19.5]
+********
+Updated
+-------
+* Utilized refactored event manager in central node from v2 in ska-tmc-common
+* ska-tmc-common version updated to 0.27.5
+
+[0.19.4]
+********
+* Utilized the latest common version to 0.27.4
+* Added minor changes to loggers.
+
 [0.19.3]
 ********
 * Updated the common version to 0.27.2
@@ -11,10 +24,6 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 * Fixed minor bugs in logs on telescope_off_command and telescope_standby_command involving incorrect number of variables provided to parametrized logs.
 * Implemented string typecasting in lazy logging to deal with the logging errors involving MemoryError.
 
-[0.19.4]
-********
-* Utilized the latest common version to 0.27.4
-* Added minor changes to loggers.
 
 [0.19.2]
 ********